requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
django-filter==25.1
djangorestframework-simplejwt==5.4.0
pydash==8.0.5